+++ src/libs/constants/constants.ts
@@ -1,4 +1,6 @@
+export { APPLICATION_NAME } from "./application-name.constant.js";
 export { FIRST_ARRAY_ELEMENT } from "./first-array-element.constant.js";
+export { LAST_ARRAY_ELEMENT } from "./last-array-element.constant.js";
 export {
 	NUMBER_OF_REVIEWS_TO_RENDER,
 	REVIEWS,