+++ src/libs/constants/last-array-element.constant.ts
@@ -1,0 +1,3 @@
+const LAST_ARRAY_ELEMENT = -1;
+
+export { LAST_ARRAY_ELEMENT };