+++ src/libs/modules/random-words/base-random-words.module.ts
@@ -1,0 +1,11 @@
+import { generate } from "random-words";
+
+import { type RandomWords } from "./libs/types/types.js";
+
+class BaseRandomWords implements RandomWords {
+	public generateWord(): string {
+		return generate() as string;
+	}
+}
+
+export { BaseRandomWords };