+++ src/libs/modules/task-scheduler/task-scheduler.ts
@@ -1,0 +1,7 @@
+import { BaseTaskScheduler } from "./base-task-scheduler.module.js";
+
+const taskScheduler = new BaseTaskScheduler();
+
+export { taskScheduler };
+export { CronExpression } from "./libs/enums/enums.js";
+export { type TaskScheduler } from "./libs/types/types.js";