+++ src/middlewares/inject-data/inject-data.middleware.ts
@@ -1,5 +1,6 @@
 import { type NextFunction, type Request, type Response } from "express";
 
+import { asyncHandler } from "~/libs/helpers/helpers.js";
 import { type Application, type Middleware } from "~/libs/types/types.js";
 import { type AuthService } from "~/modules/auth/auth.js";
 
@@ -11,13 +12,15 @@ class InjectDataMiddleware implements Middleware {
 	}
 
 	public init(app: Application): void {
-		app.use(async (req: Request, _: Response, next: NextFunction) => {
-			if (req.cookies?.token) {
-				req.user = await this.authService.findByToken(req.cookies?.token);
-			}
+		app.use(
+			asyncHandler(async (req: Request, _: Response, next: NextFunction) => {
+				if (req.cookies?.token) {
+					req.user = await this.authService.findByToken(req.cookies?.token);
+				}
 
-			next();
-		});
+				next();
+			}),
+		);
 	}
 
 	public get name(): string {